vizable_back/requirements.txt
altair==4.2.0
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
blinker==1.5
boto3==1.26.11
botocore==1.29.11
cachetools==5.2.0
certifi==2022.9.24
cftime==1.6.2
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cloudpickle==2.2.0
commonmark==0.9.1
contourpy==1.0.5
customtkinter==4.6.3
cycler==0.11.0
darkdetect==0.7.1
dask==2022.10.2
debugpy==1.6.3
decorator==5.1.1
Django==4.1.3
djangorestframework==3.14.0
entrypoints==0.4
exceptiongroup==1.0.1
executing==1.0.0
fonttools==4.37.2
fsspec==2022.10.0
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
imageio==2.22.2
importlib-metadata==5.0.0
iniconfig==1.1.1
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
netCDF4==1.6.1
networkx==2.8.7
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
partd==1.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
portalocker==2.6.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.0
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pysat==3.0.5
pytest==7.2.0
python-dateutil==2.8.2
python-sat==0.1.7.dev19
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pywatchman==1.4.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.0
regex==2022.9.13
requests==2.28.1
rich==12.6.0
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.9.2
seaborn==0.12.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
stack-data==0.5.0
streamlit==1.14.1
tifffile==2022.10.10
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wcwidth==0.2.5
xarray==2022.11.0
zipp==3.10.0
